Remove /tests by moving remaining runtime tests into /runtime/test
