+++ pom.xml
@@ -67,7 +67,7 @@ limitations under the License.
         <module>runtime/master</module>
         <module>runtime/driver</module>
         <module>runtime/plangenerator</module>
-        <module>tests</module>
+        <module>runtime/test</module>
     </modules>
 
     <dependencies>
