+++ runtime/dag/plan-logical.json
@@ -1,0 +1,1 @@
+{"vertices": [{"id": "Stage-0", "properties": {"scheduleGroup": 0, "irDag": {"vertices": [{"id": "vertex-1", "properties": {"class": "OperatorVertex", "executionProperties": {"edu.snu.nemo.common.ir.vertex.executionproperty.ExecutorPlacementProperty": "None", "edu.snu.nemo.common.ir.vertex.executionproperty.ParallelismProperty": "5", "edu.snu.nemo.common.ir.vertex.executionproperty.ScheduleGroupProperty": "0"}, "transform": "edu.snu.nemo.common.test.EmptyComponents$EmptyTransform@1877ab81, name: "}}], "edges": []}, "parallelism": 5, "executionProperties": {"edu.snu.nemo.common.ir.vertex.executionproperty.ExecutorPlacementProperty": "None", "edu.snu.nemo.common.ir.vertex.executionproperty.ParallelismProperty": "5", "edu.snu.nemo.common.ir.vertex.executionproperty.ScheduleGroupProperty": "0"}}}, {"id": "Stage-1", "properties": {"scheduleGroup": 1, "irDag": {"vertices": [{"id": "vertex-2", "properties": {"class": "OperatorVertex", "executionProperties": {"edu.snu.nemo.common.ir.vertex.executionproperty.ExecutorPlacementProperty": "None", "edu.snu.nemo.common.ir.vertex.executionproperty.ParallelismProperty": "3", "edu.snu.nemo.common.ir.vertex.executionproperty.ScheduleGroupProperty": "1"}, "transform": "edu.snu.nemo.common.test.EmptyComponents$EmptyTransform@1877ab81, name: "}}], "edges": []}, "parallelism": 3, "executionProperties": {"edu.snu.nemo.common.ir.vertex.executionproperty.ExecutorPlacementProperty": "None", "edu.snu.nemo.common.ir.vertex.executionproperty.ParallelismProperty": "3", "edu.snu.nemo.common.ir.vertex.executionproperty.ScheduleGroupProperty": "1"}}}], "edges": [{"src": "Stage-0", "dst": "Stage-1", "properties": {"runtimeEdgeId": "edge-1", "executionProperties": {"edu.snu.nemo.common.ir.edge.executionproperty.PartitionerProperty": "IntactPartitioner", "edu.snu.nemo.common.ir.edge.executionproperty.DataCommunicationPatternProperty": "OneToOne", "edu.snu.nemo.common.ir.edge.executionproperty.DataFlowModelProperty": "Pull", "edu.snu.nemo.common.ir.edge.executionproperty.InterTaskDataStoreProperty": "MemoryStore"}, "externalSrcVertexId": "vertex-1", "externalDstVertexId": "vertex-2"}}]}
